Add files via upload
WebShell is a collection of tools developed for cybersecurity professionals and penetration testers that allow executing commands on target systems using server-side scripting languages. It includes a variety of web shell applications for PHP, ASP.NET, and other web technologies. These tools are used to detect vulnerabilities, test the robustness of systems, and perform security assessments. It is intended for legal and ethical penetration testing only.

#WebShell #SiberGüvenlik #PenetrationTesting #HackingTools #PHP #WebExploitation #SecurityTesting #EthicalHacking #BackdoorTools #CommandExecution